refactor(dashboard): extract job DAG layout into pure module
