+++ dashboard/src/features/jobs/components/job-dag-tab.tsx
@@ -2,8 +2,9 @@ import { Link } from "@tanstack/react-router";
 import { Workflow } from "lucide-react";
 import { useMemo } from "react";
 import { EmptyState, ErrorState, Skeleton } from "@/components/ui";
-import type { DagData, DagEdge, DagNode, JobStatus } from "@/lib/api-types";
+import type { DagData, DagEdge, JobStatus } from "@/lib/api-types";
 import { JOB_STATUS_LABEL } from "@/lib/status";
+import { type LaidOutNode, layout, NODE_HEIGHT, NODE_WIDTH } from "./dag-layout";
 
 interface JobDagTabProps {
   dag: DagData | undefined;
@@ -29,99 +30,6 @@ const STATUS_STROKE: Record<JobStatus, string> = {
   dead: "var(--color-danger)",
   cancelled: "var(--color-warning)",
 };
-
-const NODE_WIDTH = 180;
-const NODE_HEIGHT = 58;
-const LAYER_GAP_X = 80;
-const NODE_GAP_Y = 24;
-const PADDING = 32;
-
-interface LaidOutNode extends DagNode {
-  x: number;
-  y: number;
-  layer: number;
-}
-
-/**
- * Lay out DAG nodes in BFS layers from sources (nodes with no inbound edges).
- * Each layer's nodes are centered vertically so the graph is readable even
- * for lopsided fan-outs. Returns positioned nodes plus canvas dimensions.
- */
-function layout(data: DagData): { nodes: LaidOutNode[]; width: number; height: number } {
-  const nodes = data.nodes;
-  const edges = data.edges;
-  if (nodes.length === 0) return { nodes: [], width: 0, height: 0 };
-
-  const incoming = new Map<string, number>();
-  const outgoing = new Map<string, string[]>();
-  for (const n of nodes) incoming.set(n.id, 0);
-  for (const e of edges) {
-    incoming.set(e.to, (incoming.get(e.to) ?? 0) + 1);
-    const arr = outgoing.get(e.from);
-    if (arr) arr.push(e.to);
-    else outgoing.set(e.from, [e.to]);
-  }
-
-  const layerOf = new Map<string, number>();
-  const queue: string[] = [];
-  for (const n of nodes) {
-    if ((incoming.get(n.id) ?? 0) === 0) {
-      layerOf.set(n.id, 0);
-      queue.push(n.id);
-    }
-  }
-  // Handle cycles defensively: any node we haven't ranked sits on layer 0.
-  if (queue.length === 0 && nodes.length > 0) {
-    layerOf.set(nodes[0]!.id, 0);
-    queue.push(nodes[0]!.id);
-  }
-
-  while (queue.length > 0) {
-    const id = queue.shift()!;
-    const depth = layerOf.get(id) ?? 0;
-    for (const child of outgoing.get(id) ?? []) {
-      const nextDepth = depth + 1;
-      if ((layerOf.get(child) ?? -1) < nextDepth) {
-        layerOf.set(child, nextDepth);
-        queue.push(child);
-      }
-    }
-  }
-
-  const layers = new Map<number, string[]>();
-  for (const n of nodes) {
-    const layer = layerOf.get(n.id) ?? 0;
-    const bucket = layers.get(layer) ?? [];
-    bucket.push(n.id);
-    layers.set(layer, bucket);
-  }
-
-  const sortedLayers = [...layers.entries()].sort(([a], [b]) => a - b);
-  const tallestLayerSize = sortedLayers.reduce((max, [, ids]) => Math.max(max, ids.length), 0);
-  const canvasHeight =
-    PADDING * 2 + tallestLayerSize * NODE_HEIGHT + (tallestLayerSize - 1) * NODE_GAP_Y;
-  const canvasWidth =
-    PADDING * 2 + sortedLayers.length * NODE_WIDTH + (sortedLayers.length - 1) * LAYER_GAP_X;
-
-  const laidOut: LaidOutNode[] = [];
-  const byId = new Map(nodes.map((n) => [n.id, n]));
-  for (const [layer, ids] of sortedLayers) {
-    const layerHeight = ids.length * NODE_HEIGHT + (ids.length - 1) * NODE_GAP_Y;
-    const startY = (canvasHeight - layerHeight) / 2;
-    ids.forEach((id, i) => {
-      const base = byId.get(id);
-      if (!base) return;
-      laidOut.push({
-        ...base,
-        layer,
-        x: PADDING + layer * (NODE_WIDTH + LAYER_GAP_X),
-        y: startY + i * (NODE_HEIGHT + NODE_GAP_Y),
-      });
-    });
-  }
-
-  return { nodes: laidOut, width: canvasWidth, height: canvasHeight };
-}
 
 export function JobDagTab({ dag, loading, error, onRetry }: JobDagTabProps) {
   const layoutResult = useMemo(
